#200: Created Selection class, moved all unused selection methods to test scope.
