+++ core/src/main/java/io/parsingdata/metal/data/ImmutableList.java
@@ -19,7 +19,7 @@ package io.parsingdata.metal.data;
 import static io.parsingdata.metal.SafeTrampoline.complete;
 import static io.parsingdata.metal.SafeTrampoline.intermediate;
 import static io.parsingdata.metal.Util.checkNotNull;
-import static io.parsingdata.metal.data.transformation.Reversal.reverse;
+import static io.parsingdata.metal.data.Selection.reverse;
 
 import java.util.Objects;
