+++ core/src/test/java/io/parsingdata/metal/ImmutableListTest.java
@@ -23,7 +23,7 @@ import static org.junit.Assert.assertTrue;
 import static io.parsingdata.metal.Shorthand.def;
 import static io.parsingdata.metal.data.selection.ByName.get;
 import static io.parsingdata.metal.data.selection.ByName.getAll;
-import static io.parsingdata.metal.data.transformation.Reversal.reverse;
+import static io.parsingdata.metal.data.Selection.reverse;
 import static io.parsingdata.metal.Util.createFromBytes;
 import static io.parsingdata.metal.util.EncodingFactory.enc;
