+++ core/src/test/java/io/parsingdata/metal/data/selection/ByType.java
